requirements.txt
dj-database-url==0.5.0
Django==2.0.5
django-cors-headers==2.2.0
django-dotenv==1.4.2
django-environ==0.4.5
django-heroku==0.3.1
django-model-utils==3.1.2
djangorestframework==3.8.2
gunicorn==19.9.0
psycopg2==2.8.2
psycopg2-binary==2.8.2
pytz==2019.1
whitenoise==4.1.2
django-reversion
